requirements.txt
requests>=2.12.5
click>=7.0
PyYAML>=3.13
orjson>=3.0.0
pygments>=2.5.0
tqdm>=4.0.0